sry got cut off
Tried to add funny images. But i don't want to deal with the css.
So I'm just going to add a way to input a name. A button to restart
the game. A button that starts the game, it just doesn't do anything.
And a display element that is inputted at the top saying who won.
Then i'll call this project done. It's just extra stuff to finish