Finished requirements
Since this isn't a capstone project I didn't do any css.
While only focusing on the fundamental oop. And moving on to harder
projects like the capstone that will be add to the portfolio.
I know as a beginner dev I have much to learn so might as well move
fast and learn more instead of making everything perfect.